* release/1.3.0: (49 commits)
  Use BufferList
  FDB-261 reverted version number
  FDB-261 odc header handle EOF
  Use the correct destruction function in the fortran test
  Use the correct destruction function in the fortran test
  ODB-517: Remove the DataHandleFactory and all HttpHandle remnants
  ODB-517: Remove dependence on StdFileHandle (c.f. ECKIT-547)
  ODB-517: Remove the DataHandleFactory and all HttpHandle remnants
  ODB-517: Remove dependence on StdFileHandle (c.f. ECKIT-547)
  Make the undefined behaviour sanitizer happy
  Make the undefined behaviour sanitizer happy
  Fix split test for comparator change
  Fix split test for comparator change
  Work around cray compiler bug
  Reintroduce comparator hack to handle ODB migrator iterator with incomplete knowledge of missing values
  Fix fortran test
  Fix fortran test
  version
  Don't allocate/throw away PathName objects in CompareTool
  Correct destructor on encode buffer in odc.cc
  ...

+++ src/odc/codec/Integer.h
@@ -73,8 +73,11 @@ private: // methods
     unsigned char* encode(unsigned char* p, const double& d) override {
         static_assert(sizeof(ValueType) == sizeof(d), "unsafe casting check");
 
+        // Static cast to uint64_t ensures that the narrowing cast doesn't trigger the
+        // undefined behaviour sanitizer. It is correct that we can generate an integer
+        // that is too big, and throw it away, if invalid data gets here.
         const ValueType& val(reinterpret_cast<const ValueType&>(d));
-        InternalValueType s = val - this->min_;
+        InternalValueType s = static_cast<uint64_t>(val - this->min_);
         ByteOrder::swap(s);
         ::memcpy(p, &s, sizeof(s));
         return p + sizeof(s);
